requirements.txt
matplotlib==3.5.1
pandas==1.4.2
Pillow==9.0.1
Pillow==10.2.0
plotly==5.6.0
streamlit==1.31.1
